cloud9-solution/requirements.txt
google-genai>=1.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
